Fix output of JsonIgnore attributes
